Added options, added timeout, renamed history class
Added options so that functions which call willModify() can have
the ability to not call it. This is only done when receiving previous
data.

Added timeout so that all undo states that occur within a time (350ms)
are condensed into a single state (specifically the latest
chronologically) to save space and undo time.

Renamed class History and history variable to avoid confusion with
window.history